factored out raw input value extraction
