+++ extensions-core/datasketches/src/main/java/org/apache/druid/query/aggregation/datasketches/theta/SketchBuildComplexMetricSerde.java
@@ -19,7 +19,7 @@
 
 package org.apache.druid.query.aggregation.datasketches.theta;
 
-import org.apache.druid.data.input.InputRow;
+import org.apache.druid.query.aggregation.datasketches.RawInputValueExtractor;
 import org.apache.druid.segment.serde.ComplexMetricExtractor;
 
 /**
@@ -29,20 +29,6 @@ public class SketchBuildComplexMetricSerde extends SketchMergeComplexMetricSerde
   @Override
   public ComplexMetricExtractor getExtractor()
   {
-    return new ComplexMetricExtractor()
-    {
-
-      @Override
-      public Class<?> extractedClass()
-      {
-        return Object.class;
-      }
-
-      @Override
-      public Object extractValue(InputRow inputRow, String metricName)
-      {
-        return inputRow.getRaw(metricName);
-      }
-    };
+    return new RawInputValueExtractor();
   }
 }
